requirements.txt
aioftp==0.20.1
aiohttp==3.8.1
aiosignal==1.2.0
astropy==5.0.1
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
multidict==6.0.2
numpy==1.22.2
packaging==21.3
parfive==1.5.1
pyerfa==2.0.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
sunpy==3.1.4
tqdm==4.63.0
Werkzeug==2.0.3
yarl==1.7.2
